Add OpenMP support and refactor options
OpenMP support has been added and the build options have been refactored in CMakeLists.txt. Now, both multiprecision and OpenMP are optional and can be toggled on and off with ENABLE_MULTIPRECISION and ENABLE_OPENMP options, respectively. Also, a parallel reduction has been applied to a triple nested loop in helper_functions.cpp to leverage OpenMP. These changes provide users with more flexibility and can potentially improve the performance of the application when OpenMP is enabled.